requirements.txt
streamlit
tensorflow
numpy
pillow
google-generativeai
